pyproject.toml
[tool.poetry]
name = "servo-jenkins-trigger"
version = "0.1.0"
description = "Opsani Servo plugin for triggering Jenkins (e.g load pipeline)"
authors = ["Robert Starmer <robert@opsani.com>"]
license = "BSD-3-Clause"

packages = [
    { include = "measure" },
]
include = ["README.md", "LICENSE"]

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3"
requests = "^2.23"
python-jenkins = "^1.7"
jenkinsapi = "^0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"